trained_model/requirements.txt
Flask
pandas
numpy
scikit-learn
gunicorn
mysql-connector-python